Dao: fix notifications
